// src/app/enums/Stock.ts
export enum EStockNavigation {
  Goods = 'goods',
  Stock = 'stock',
  Warehouses = 'warehouses',
  TechCard = 'techcard',
  Tasks = 'tasks',
  Counterparties = 'counterparties'
}

export enum EStockOperations {
  Acceptance = 'acceptance',
  Capitalization = 'capitalization',
  Movement = 'movement',
  WriteOff = 'write-off'
}